고영희 생성 / 수정 / 상세 기능 추가

+++ app/src/main/java/com/example/dongnaegoyang/data/remote/model/response/CatDetailInfoResponse.kt
@@ -53,7 +53,7 @@ data class CatDetailInfoResponse(
 )
 
 data class PhotoList(
-    @SerializedName("imageIdx") val imageIdx: Int,
+    @SerializedName("imageIdx") val imageIdx: Long,
     @SerializedName("url") val url: String
 ): Serializable
